Sphinx documentation and tutorial notebook for T1 (#62)
* sphinx documentation for T1

* sphinx fix

* started the t1 tutorial

* basic code in tutorial

* expanded the tutorial

* fix import

* fixed tutorial

* make the tests pass

* Rewrote docstring ot T1Analysis

* lint

* sphinx fixes

* sphinx fixes

* Removed utf encoding (to make verify_headers happy)

* removed connecting line when in error bars plotting

* fixes

* replaced C by B

* removed execution step in t1 doc string
